Remove ra_fmt crate
