--- crates/ra_fmt/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-edition = "2018"
-name = "ra_fmt"
-version = "0.1.0"
-authors = ["rust-analyzer developers"]
-publish = false
-license = "MIT OR Apache-2.0"
-
-[lib]
-doctest = false
-
-[dependencies]
-itertools = "0.9.0"
-
-syntax = { path = "../syntax" }
