+++ crates/ra_ide/Cargo.toml
@@ -26,7 +26,6 @@ text_edit = { path = "../text_edit" }
 ra_db = { path = "../ra_db" }
 ra_ide_db = { path = "../ra_ide_db" }
 cfg = { path = "../cfg" }
-ra_fmt = { path = "../ra_fmt" }
 profile = { path = "../profile" }
 test_utils = { path = "../test_utils" }
 ra_assists = { path = "../ra_assists" }
